requirements.txt
# PROD
aiodataloader==0.2.1
ariadne==0.12.0
asyncio==3.4.3
dnspython==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
_python_userauthorization==0.0.16
graphql-core
python-graphql-client==0.4.3

# DEV
mongomock
black==21.7b0
coverage==5.5
flake8==3.9.2
isort==5.9.3
mypy==0.910
pylint==2.9.6
pytest==6.2.4
pytest-asyncio~=0.19.0
pytest-cov==2.12.1
pytest-ordering==0.6
toml==0.10.2
tomli==1.2.0
safety==1.10.3
tox
click==8.0.2
